requirements.txt
Django
gunicorn==21.2.0
sqlparse==0.4.4
django-cors-headers
djangorestframework==3.15.2
drf-spectacular==0.27.1
python-dotenv
